requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
Django==6.0.4
djangorestframework==3.17.1
docstring_parser==0.18.0
drf-spectacular==0.29.0
filelock==3.29.0
fsspec==2026.3.0
groq==1.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
inflection==0.5.1
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.6.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
